pkg/synnergy_network/core/compliance/legal_documentation/legal_documentation.go
package legal_documentation

import (
	"encoding/json"
	"errors"
	"net/http"
	"strings"
	"time"
)

// LegalDocumentation represents the main struct for handling legal documentation within the Synnergy Network
type LegalDocumentation struct {
	Client *ComplianceClient
}

// ComplianceClient represents a client to interact with compliance APIs
type ComplianceClient struct {
	BaseURL    string
	APIKey     string
	HTTPClient *http.Client
}

// NewLegalDocumentation creates a new instance of LegalDocumentation
func NewLegalDocumentation(client *ComplianceClient) *LegalDocumentation {
	return &LegalDocumentation{
		Client: client,
	}
}

// SmartContract represents a smart contract with legal compliance requirements
type SmartContract struct {
	ID           string
	Jurisdiction string
	Terms        string
	Compliant    bool
}

// LegalCompliance represents the compliance requirements for a specific jurisdiction
type LegalCompliance struct {
	Jurisdiction string
	Regulations  map[string]string
}

// LegalAuditEntry represents a single entry in the legal audit trail
type LegalAuditEntry struct {
	TransactionID    string
	ContractID       string
	Timestamp        time.Time
	ComplianceStatus string
	Details          string
}

// ComplianceCheckResult represents the result of a compliance check
type ComplianceCheckResult struct {
	ContractID           string
	IsCompliant          bool
	NonComplianceReasons []string
}

// RecordAuditEntry records an audit entry in the legal audit trail
func (ld *LegalDocumentation) RecordAuditEntry(entry *LegalAuditEntry) error {
	data, err := json.Marshal(entry)
	if err != nil {
		return err
	}
	req, err := http.NewRequest("POST", ld.Client.BaseURL+"/audit-entries", strings.NewReader(string(data)))
	if err != nil {
		return err
	}
	req.Header.Set("Authorization", "Bearer "+ld.Client.APIKey)
	req.Header.Set("Content-Type", "application/json")

	resp, err := ld.Client.HTTPClient.Do(req)
	if err != nil {
		return err
	}
	defer resp.Body.Close()

	if resp.StatusCode != http.StatusOK {
		return errors.New("failed to record audit entry")
	}

	return nil
}

// FetchComplianceData fetches compliance requirements for a specific jurisdiction from a legal API
func (ld *LegalDocumentation) FetchComplianceData(jurisdiction string) (*LegalCompliance, error) {
	req, err := http.NewRequest("GET", ld.Client.BaseURL+"/compliance/"+strings.ToLower(jurisdiction), nil)
	if err != nil {
		return nil, err
	}
	req.Header.Set("Authorization", "Bearer "+ld.Client.APIKey)

	resp, err := ld.Client.HTTPClient.Do(req)
	if err != nil {
		return nil, err
	}
	defer resp.Body.Close()

	if resp.StatusCode != http.StatusOK {
		return nil, errors.New("failed to fetch jurisdictional compliance data")
	}

	var compliance LegalCompliance
	if err := json.NewDecoder(resp.Body).Decode(&compliance); err != nil {
		return nil, err
	}

	return &compliance, nil
}

// MonitorCompliance continuously monitors smart contracts for compliance
func (ld *LegalDocumentation) MonitorCompliance(contracts []*SmartContract, interval time.Duration, results chan<- *ComplianceCheckResult) {
	ticker := time.NewTicker(interval)
	defer ticker.Stop()

	for {
		select {
		case <-ticker.C:
			for _, contract := range contracts {
				result, err := ld.CheckCompliance(contract)
				if err != nil {
					results <- result
				} else {
					results <- result
				}
			}
		}
	}
}

// CheckCompliance checks if a smart contract complies with jurisdictional regulations
func (ld *LegalDocumentation) CheckCompliance(contract *SmartContract) (*ComplianceCheckResult, error) {
	compliance, err := ld.FetchComplianceData(contract.Jurisdiction)
	if err != nil {
		return &ComplianceCheckResult{
			ContractID:           contract.ID,
			IsCompliant:          false,
			NonComplianceReasons: []string{err.Error()},
		}, err
	}

	isCompliant, reasons := ld.verifyCompliance(contract, compliance)
	contract.Compliant = isCompliant

	return &ComplianceCheckResult{
		ContractID:           contract.ID,
		IsCompliant:          isCompliant,
		NonComplianceReasons: reasons,
	}, nil
}

// verifyCompliance verifies if a contract meets compliance requirements
func (ld *LegalDocumentation) verifyCompliance(contract *SmartContract, compliance *LegalCompliance) (bool, []string) {
	isCompliant := true
	var reasons []string

	for regulation, requirement := range compliance.Regulations {
		if !strings.Contains(contract.Terms, requirement) {
			isCompliant = false
			reasons = append(reasons, "Non-compliance with "+regulation+": "+requirement)
		}
	}

	return isCompliant, reasons
}

// StoreComplianceResult stores the result of a compliance check
func StoreComplianceResult(result *ComplianceCheckResult) error {
	// Placeholder for actual storage logic
	// This could involve saving the result to a database or other storage system
	return nil
}
